// frontend/src/enums/dataQuery.enum.ts
export enum DataQueryEnum {
  GET = "GET",
  POST = "POST",
  PUT = "PUT",
  DELETE = "DELETE",
  INVALID_REFRESH_TOKEN = "INVALID_REFRESH_TOKEN",
  UNAUTHORIZED = "UNAUTHORIZED",
  FORBIDDEN = "FORBIDDEN",
  NOT_FOUND = "NOT_FOUND",
  FAILED_TO_CREATE_INVOICE = "FAILED_TO_CREATE_INVOICE",
  SUCCESS = "SUCCESS",
}